Refactor types from profile snapshot
