+++ src/components/EditProfile/PersonasSection.tsx
@@ -5,11 +5,7 @@ import { Textarea } from "@/components/ui/textarea"
 import { Button } from "@/components/ui/button"
 import { Edit, Plus, X } from "lucide-react"
 import clsx from "clsx"
-
-export interface Persona {
-  title: string
-  bullets: string[]
-}
+import type { Persona } from "@/types/profile"
 
 interface PersonaEditState {
   title: string
